+++ R/pirateplot_function.R
@@ -2,7 +2,7 @@
 #'
 #' The pirateplot function creates an RDI (Raw data, Descriptive and Inferential statistic) plot showing the relationship between 1 to 3 categorical independent variables and 1 continuous dependent variable.
 #'
-#' @param formula formula. A formula in the form \code{y ~ x1 + x2 + x3} indicating the vertical response variable (y) and up to three independent varaibles
+#' @param formula formula. A formula in the form \code{y ~ x1 + x2 + x3} indicating the vertical response variable (y) and up to three independent variables
 #' @param data dataframe. A dataframe containing the variables specified in formula.
 #' @param plot logical. If \code{TRUE} (the default), thent the pirateplot is produced. If \code{FALSE}, the data summaries created in the plot are returned as a list.
 #' @param pal string. The color palette of the plot. Can be a single color, a vector of colors, or the name of a palette in the piratepal() function (e.g.; "basel", "google", "southpark"). To see all the palettes, run \code{piratepal(palette = "all", action = "show")}
@@ -15,29 +15,31 @@
 #' @param bean.lwd,bean.lty,inf.lwd,avg.line.lwd,bar.lwd numeric. Vectors of numbers customizing the look of beans and lines.
 #' @param width.min,width.max numeric. The minimum/maximum width of the beans.
 #' @param cut.min,cut.max numeric. Optional minimum and maximum values of the beans.
-#' @param inf string. A string indicating what types of inference bands to calculate. "ci" means frequentist confidence intervals, "hdi" means Bayesian Highest Density Intervals (HDI), "iqr" means interquartile range.
-#' @param inf.band string. Either \code{"wide"} to indicate a fixed width band, or \code{"tight"} to indicate a band constrained by the bean
-#' @param inf.p numeric. A number between 0 and 1 indicating the level of confidence to use in calculating inferences for either confidence intervals or HDIs. The default is 0.95
+#' @param inf.method string. A string indicating what types of inference bands to calculate. "ci" means frequentist confidence intervals, "hdi" means Bayesian Highest Density Intervals (HDI), "iqr" means interquartile range, "sd" means standard deviation, "se" means standard error, "withinci" means frequentist confidence intervals in a within design (Morey, 2008).
+#' @param inf.within string. The variable which serves as an ID variable in a within design.
+#' @param inf.disp string. How should inference ranges be displayed? \code{"line"} creates a classic vertical line, \code{"rect"} creates a rectangle, \code{"bean"} forms the inference around the bean.
+#' @param inf.p numeric. A number adjusting how inference ranges are calculated. for \code{"ci"} and \code{"hdi"}, a number between 0 and 1 indicating the level of confidence (default is .95). For \code{"sd"} and \code{"se"}, the number of standard deviations / standard errors added to or subtracted from the mean (default is 1).
 #' @param hdi.iter integer. Number of iterations to run when calculating the HDI. Larger values lead to better estimates, but can be more time consuming.
 #' @param bw,adjust Arguments passed to density calculations for beans (see ?density)
 #' @param jitter.val numeric. Amount of jitter added to points horizontally. Defaults to 0.05.
-#' @param at integer. Locations of the beans. Especially helpful when adding beans to an existing plot with add = T
+#' @param at integer. Locations of the beans. Especially helpful when adding beans to an existing plot with add = TRUE
 #' @param sortx string. How to sort the x values. Can be "sequential" (as they are found in the original dataframe), "alphabetical", or a string indicating a function (i.e.; "mean")
 #' @param add logical. Whether to add the pirateplot to an existing plotting space or not.
 #' @param evidence logical. Should Bayesian evidence be shown? (currently ignored)
+#' @param cap.beans logical. Should maximum and minimum values of the bean densities be capped at the limits found in the data? Default is FALSE.
 #' @param quant.length,quant.lwd numeric. Specifies line lengths/widths of \code{quant}.
+#' @param quant.boxplot logical. Should standard values be included?
 #' @param family a font family (Not currently in use)
 #' @param cex.lab,cex.axis Size of the labels and axes.
 #' @param gl.lwd,gl.lty Customization for grid lines.
 #' @param bty,xlim,ylim,xlab,ylab,main,yaxt,xaxt General plotting arguments
 #' @param quant numeric. Adds horizontal lines representing custom quantiles.
-#' @param bar.b.lwd,line.fun,inf.o,bean.o,inf.col,theme.o depricated arguments
-#' @param ... other arguments passed on to the plot function (e.g.; main, xlab, ylab, ylim, cex.axis, cex.main, cex.lab)
+#' @param bar.b.lwd,line.fun,inf.o,bean.o,inf.col,theme.o,inf,inf.type,inf.band,bar.o,line.o,hdi.o depricated arguments
 #' @keywords plot
 #' @importFrom BayesFactor ttestBF
 #' @importFrom grDevices col2rgb gray rgb
 #' @importFrom graphics abline axis layout mtext par plot points polygon rasterImage rect segments text
-#' @importFrom stats density model.frame optimize rnorm t.test qbeta sd quantile
+#' @importFrom stats density model.frame optimize rnorm t.test qbeta sd quantile IQR
 #' @importFrom utils vignette
 #' @export
 #' @examples
@@ -58,6 +60,18 @@
 #'           data = ChickWeight,
 #'           main = "Chicken weight by time",
 #'           theme = 2) # theme 2
+#'
+#'# theme 3
+#'pirateplot(formula = weight ~ Time,
+#'           data = ChickWeight,
+#'           main = "Chicken weight by time",
+#'           theme = 3) # theme 3
+#'
+#'# theme 4
+#'pirateplot(formula = weight ~ Time,
+#'           data = ChickWeight,
+#'           main = "Chicken weight by time",
+#'           theme = 4) # theme 4
 #'
 #'# Start with theme 2, but then customise!
 #'pirateplot(formula = weight ~ Time,
@@ -138,7 +152,7 @@ pirateplot <- function(
   avg.line.lwd = 4,
   bean.lwd = 1,
   bean.lty = 1,
-  inf.lwd = 1,
+  inf.lwd = NULL,
   bar.lwd = 1,
   at = NULL,
   bw = "nrd0",
@@ -150,13 +164,16 @@ pirateplot <- function(
   quant = NULL,
   quant.length = NULL,
   quant.lwd = NULL,
+  quant.boxplot = FALSE,
   bty = "o",
   evidence = FALSE,
+  cap.beans = FALSE,
   family = NULL,
-  inf = "hdi",
-  inf.p = .95,
+  inf.method = "hdi",
+  inf.within = NULL,
+  inf.p = NULL,
   hdi.iter = 1e3,
-  inf.band = "wide",
+  inf.disp = NULL,
   cut.min = NULL,
   cut.max = NULL,
   width.min = .3,
@@ -172,18 +189,23 @@ pirateplot <- function(
   gl.lty = NULL,
   bar.b.lwd = NULL,
   line.fun = NULL,
+  line.o = NULL,
   inf.o = NULL,
   bean.o = NULL,
   inf.col = NULL,
   theme.o = NULL,
-  ...
+  bar.o = NULL,
+  inf = NULL,
+  hdi.o = NULL,
+  inf.type = NULL,
+  inf.band = NULL
 ) {
 #
 #
-  # formula = time ~ sequel + genre + rating
-  # data = subset(movies,
-  #               genre %in% c("Action", "Adventure", "Comedy", "Horror") &
-  #                 time > 0)
+#
+#
+  # formula = weight ~ Time
+  # data = ChickWeight
   # plot = TRUE
   # avg.line.fun = mean
   # pal = "basel"
@@ -215,7 +237,7 @@ pirateplot <- function(
   # avg.line.lwd = 4
   # bean.lwd = 1
   # bean.lty = 1
-  # inf.lwd = 1
+  # inf.lwd = NULL
   # bar.lwd = 1
   # at = NULL
   # bw = "nrd0"
@@ -230,10 +252,10 @@ pirateplot <- function(
   # bty = "o"
   # evidence = FALSE
   # family = NULL
-  # inf = "hdi"
+  # inf.method = "hdi"
   # inf.p = .95
   # hdi.iter = 1e3
-  # inf.band = "wide"
+  # inf.disp = "line"
   # cut.min = NULL
   # cut.max = NULL
   # width.min = .3
@@ -253,12 +275,14 @@ pirateplot <- function(
   # bean.o = NULL
   # inf.col = NULL
   # theme.o = NULL
+  # inf = NULL
+  # inf.type = NULL
+  # inf.band = NULL
   #
-  #
-  # formula = time ~ sequel + genre + rating
-  # data = subset(movies,
-  #               genre %in% c("Action", "Adventure", "Comedy", "Horror") &
-  #                 time > 0)
+  # formula = weight ~ Time
+  # data = ChickWeight
+  # theme = 2
+  # main = "theme = 2"
 
 # -----
 #  SETUP
@@ -289,6 +313,15 @@ if(is.null(inf.o) == FALSE) {
 
 }
 
+if(is.null(line.o) == FALSE) {
+
+  message("line.o is depricated. Use avg.line.o instead")
+
+  avg.line.o <- line.o
+
+}
+
+
 if(is.null(bean.o) == FALSE) {
 
   message("bean.o is depricated. Use bean.b.o instead")
@@ -312,6 +345,33 @@ if(is.null(theme.o) == FALSE) {
   theme <- theme.o
 
 }
+
+if(is.null(inf) == FALSE) {
+
+  message("inf is depricated. Use inf.method instead")
+
+  inf.method <- inf
+
+}
+
+if(is.null(inf.band) == FALSE) {
+
+  message("inf.band is depricated. Use inf.disp instead")
+
+  inf.disp <- inf.band
+
+}
+
+if(is.null(bar.o) == FALSE) {
+
+  message("bar.o is depricated. Use bar.f.o (for filling), and bar.b.o (for border) instead")
+
+  bar.f.o <- bar.o
+
+}
+
+
+
 }
 
 # Look for missing critical inputs
@@ -320,8 +380,25 @@ if(is.null(data)) {stop("You must specify a dataframe in the data argument!")}
 if(is.null(formula) | class(formula) != "formula") {stop("You must specify a valid formula in the formula argument!")}
 }
 
+# Set some defaults
+
+if(inf.method %in% c("hdi", "ci", "withinci") & is.null(inf.p)) {
+
+  inf.p <- .95
+
+}
+
+if(inf.method %in% c("sd", "se") & is.null(inf.p)) {
+
+  inf.p <- 1
+
+}
+
+
 # Reshape dataframe to include relevant variables
 {
+
+withindata <- data
 data <- model.frame(formula = formula,
                       data = data)
 
@@ -454,7 +531,7 @@ for(bean.i in 1:n.beans) {
 
   if(length(setdiff(dv.i, c(0, 1))) == 0) {
 
-    if(inf == "hdi") {
+    if(inf.method == "hdi") {
 
       # Calculate HDI from beta(Success + 1, Failure + 1)
       inf.lb <- qbeta(.025, shape1 = sum(dv.i) + 1, shape2 = sum(dv.i == 0) + 1)
@@ -462,7 +539,7 @@ for(bean.i in 1:n.beans) {
 
     }
 
-    if(inf == "ci") {
+    if(inf.method == "ci") {
 
       # Calculate 95% CI with Normal distribution approximation to binomial
       inf.lb <- mean(dv.i) - 1.96 * sqrt(mean(dv.i) * (1 - mean(dv.i)) / length(dv.i)) - .5 / length(dv.i)
@@ -472,12 +549,33 @@ for(bean.i in 1:n.beans) {
       if(inf.lb > 1) {inf.ub <- 1}
 
     }
+
+    if(inf.method == "sd") {
+
+      inf.lb <- mean(dv.i) - sd(dv.i) * inf.p
+      inf.ub <- mean(dv.i) + sd(dv.i) * inf.p
+
+      if(inf.lb < 0) {inf.lb <- 0}
+      if(inf.lb > 1) {inf.ub <- 1}
+
+    }
+
+    if(inf.method == "se") {
+
+      inf.lb <- mean(dv.i) - sd(dv.i) / sqrt(length(dv.i)) * inf.p
+      inf.ub <- mean(dv.i) + sd(dv.i) / sqrt(length(dv.i)) * inf.p
+
+      if(inf.lb < 0) {inf.lb <- 0}
+      if(inf.lb > 1) {inf.ub <- 1}
+
+    }
+
   }
 
   # Non-Binary data.i
   if(length(setdiff(dv.i, c(0, 1))) > 0) {
 
-    if(inf == "hdi") {
+    if(inf.method == "hdi") {
 
       ttest.bf <- BayesFactor::ttestBF(dv.i, posterior = T, iterations = hdi.iter, progress = F)
       samples <- ttest.bf[,1]
@@ -489,14 +587,14 @@ for(bean.i in 1:n.beans) {
 
     }
 
-    if(inf == "iqr") {
+    if(inf.method == "iqr") {
 
       inf.lb <- quantile(dv.i, probs = .25)
       inf.ub <- quantile(dv.i, probs = .75)
 
     }
 
-    if(inf == "ci") {
+    if(inf.method == "ci") {
 
       ci.i <- t.test(dv.i, conf.level = inf.p)$conf.int
 
@@ -504,6 +602,47 @@ for(bean.i in 1:n.beans) {
       inf.ub <- ci.i[2]
 
     }
+
+    if(inf.method == "sd") {
+
+      inf.lb <- mean(dv.i) - sd(dv.i) * inf.p
+      inf.ub <- mean(dv.i) + sd(dv.i) * inf.p
+
+    }
+
+    if(inf.method == "se") {
+
+      inf.lb <- mean(dv.i) - sd(dv.i) / sqrt(length(dv.i)) * inf.p
+      inf.ub <- mean(dv.i) + sd(dv.i) / sqrt(length(dv.i)) * inf.p
+
+    }
+
+    if(inf.method == "withinci") {
+
+      grandMean <- mean(dv.v)
+      Groups <- unique(withindata[inf.within])[[1]]
+      groupMean <- c()
+      dv.within <- c()
+
+      for(group in 1:length(Groups)){
+        #get the participant/group mean for each participant/group
+        groupMean[group] <- mean(dv.v[which(withindata[inf.within] == Groups[group])])
+      }
+
+      for(datum in 1:length(dv.i)){
+        #substitute group mean with grand mean, which removes between subject variance
+        dv.within[datum] <- dv.i[datum] - groupMean[datum] + grandMean
+      }
+
+      ci.i <- t.test(dv.within, conf.level = inf.p)$conf.int
+
+      ci.width <- (ci.i[2] - ci.i[1])
+      inf.lb <- mean(dv.i) - (ci.width/2) * sqrt(n.beans/(n.beans-1)) #with Morey correction
+      inf.ub <- mean(dv.i) + (ci.width/2) * sqrt(n.beans/(n.beans-1)) #with Morey correction
+
+    }
+
+
   }
 
   summary$inf.lb[bean.i] <- inf.lb
@@ -521,9 +660,9 @@ n.cols <- iv.lengths[1]
 
 # DEFINE THEMES
 {
-if((theme %in% 0:3) == FALSE) {
+if((theme %in% 0:4) == FALSE) {
 
-  print("theme must be an integer between 0 and 3. I'll set it to 1 for now.")
+  print("theme must be an integer between 0 and 4. I'll set it to 1 for now.")
   theme <- 1
 
 }
@@ -542,12 +681,16 @@ if(theme == 0) {
   if(is.null(point.cex)) {point.cex <- 1}
   if(is.null(gl.col)) {gl.col <- "white"}
 
+  if(is.null(inf.disp)) {inf.disp <- "rect"}
+
+
+
 }
 
 if(theme == 1) {
 
   if(is.null(point.o)) {point.o <- .2}
-  if(is.null(bean.b.o)) {bean.b.o <- .2}
+  if(is.null(bean.b.o)) {bean.b.o <- 1}
   if(is.null(bean.f.o)) {bean.f.o <- .2}
   if(is.null(inf.f.o)) {inf.f.o <- .8}
   if(is.null(inf.b.o)) {inf.b.o <- .8}
@@ -568,6 +711,8 @@ if(theme == 1) {
   if(is.null(point.bg)) {point.bg <- "white"}
   if(is.null(point.pch)) {point.pch <- 16}
 
+  if(is.null(inf.disp)) {inf.disp <- "rect"}
+
 
 }
 
@@ -575,7 +720,7 @@ if(theme == 2) {
 
   if(is.null(point.pch)) {point.pch <- 16}
   if(is.null(point.o)) {point.o <- .1}
-  if(is.null(bean.b.o)) {bean.b.o <- .5}
+  if(is.null(bean.b.o)) {bean.b.o <- 1}
   if(is.null(bean.f.o)) {bean.f.o <- 1}
   if(is.null(inf.f.o)) {inf.f.o <- .6}
   if(is.null(inf.b.o)) {inf.b.o <- .8}
@@ -589,15 +734,20 @@ if(theme == 2) {
   if(is.null(avg.line.col)) {avg.line.col <- "black"}
   if(is.null(bean.f.col)) {bean.f.col <- "white"}
 
+  if(is.null(inf.disp)) {inf.disp <- "rect"}
+
   if(is.null(gl.col)) {gl.col <- "gray"}
   if(is.null(gl.lwd)) {gl.lwd <- c(.5, 0)}
+
+
 
 }
 
 if(theme == 3) {
+
   if(is.null(point.pch)) {point.pch <- 16}
   if(is.null(point.o)) {point.o <- .3}
-  if(is.null(bean.b.o)) {bean.b.o <- .5}
+  if(is.null(bean.b.o)) {bean.b.o <- 1}
   if(is.null(bean.f.o)) {bean.f.o <- .5}
   if(is.null(inf.f.o))  {inf.f.o <- .9}
   if(is.null(inf.b.o)) {inf.b.o <- 1}
@@ -605,6 +755,7 @@ if(theme == 3) {
   if(is.null(bar.f.o))  {bar.f.o <- 0}
   if(is.null(bar.b.o))  {bar.b.o <- 0}
 
+  if(is.null(bean.b.col)) {bean.b.col <- "black"}
   if(is.null(inf.f.col)) {inf.f.col <- "white"}
   if(is.null(inf.b.col)) {inf.b.col <-  "black"}
   if(is.null(avg.line.col)) {avg.line.col <- "black"}
@@ -612,9 +763,54 @@ if(theme == 3) {
   if(is.null(point.cex)) {point.cex <- .5}
   #
   #  if(is.null(back.col)) {back.col <- gray(.97)}
+  if(is.null(inf.disp)) {inf.disp <- "bean"}
 
   if(is.null(gl.col)) {gl.col <- "gray"}
   if(is.null(gl.lwd)) {gl.lwd <- c(.5, 0)}
+
+
+
+}
+
+if(theme == 4) {
+
+  if(is.null(point.pch)) {point.pch <- 16}
+  if(is.null(point.o)) {point.o <- .3}
+  if(is.null(bean.b.o)) {bean.b.o <- 0}
+  if(is.null(bean.f.o)) {bean.f.o <- 0}
+  if(is.null(inf.f.o))  {inf.f.o <- 1}
+  if(is.null(inf.b.o)) {inf.b.o <- .5}
+  if(is.null(avg.line.o))  {avg.line.o <- 1}
+  if(is.null(bar.f.o))  {bar.f.o <- 1}
+  if(is.null(bar.b.o))  {bar.b.o <- 1}
+
+  if(is.null(inf.f.col)) {inf.f.col <- "black"}
+  if(is.null(inf.b.col)) {inf.b.col <-  "black"}
+  if(is.null(avg.line.col)) {avg.line.col <- "black"}
+  if(is.null(bar.f.col)) {bar.f.col <- "white"}
+  if(is.null(bar.b.col)) {bar.b.col <- "black"}
+
+  if(is.null(point.col)) {point.col <- "black"}
+  if(is.null(point.cex)) {point.cex <- .5}
+  #
+  #  if(is.null(back.col)) {back.col <- gray(.97)}
+
+  if(is.null(gl.col)) {gl.col <- "gray"}
+  if(is.null(gl.lwd)) {gl.lwd <- c(.5, 0)}
+
+  if(is.null(inf.disp)) {inf.disp <- "line"}
+
+}
+
+
+# Inference lwd depends on inf.disp...
+if(is.null(inf.lwd)) {
+
+
+  if(inf.disp == "line") {inf.lwd <- 2}
+  if(inf.disp %in% c("rect", "bean")) {inf.lwd <- 1}
+
+
 }
 
 }
@@ -689,7 +885,7 @@ if(mean(pal %in% piratepal("names")) != 1) {
 colors.df$point.col <- rep(pal, length.out = n.beans)
 colors.df$point.bg <- rep(pal, length.out = n.beans)
 colors.df$bean.b.col <- rep(pal, length.out = n.beans)
-colors.df$bean.f.col <- rep("white", length.out = n.beans)
+colors.df$bean.f.col <- rep(pal, length.out = n.beans)
 colors.df$inf.f.col <- rep(pal, length.out = n.beans)
 colors.df$inf.b.col <- rep(pal, length.out = n.beans)
 colors.df$avg.line.col <- rep(pal, length.out = n.beans)
@@ -849,8 +1045,6 @@ if(is.null(ylim) == FALSE) {
 
 }
 
-xlim <- c(min(bean.loc) - .5, max(bean.loc) + .5)
-
 # Determine x and y labels
 
 if(subplot.n.iv == 1 & is.null(xlab)) {my.xlab <- iv.names[1]}
@@ -861,8 +1055,14 @@ if(subplot.n.iv > 1) {my.xlab <- ""}
 if(is.null(ylab)) {ylab <- dv.name}
 }
 
+
+
 # PLOTTING SPACE
 if(add == FALSE) {
+
+# X-Axis
+
+if(is.null(xlim)) {xlim <- c(min(bean.loc) - .5, max(bean.loc) + .5)}
 
   plot(1,
        xlim = xlim,
@@ -887,6 +1087,8 @@ if(n.iv > 2) {
   mtext(text = top.text, side = 3, line = 1)
 
 }
+
+
 
 # Y-AXIS
 {
@@ -990,7 +1192,7 @@ if(length(dv.i) > 3) {  # only if n > 5
   dens.x.plot.i <- dens.x.i
   dens.y.plot.i <- dens.y.i
 
-  if(is.null(cut.min) == F) {
+  if(is.null(cut.min) == FALSE) {
 
     dens.x.plot.i <- dens.x.i[dens.x.i > cut.min]
     dens.y.plot.i <- dens.y.i[dens.x.i > cut.min]
@@ -998,12 +1200,21 @@ if(length(dv.i) > 3) {  # only if n > 5
   }
 
 
-  if(is.null(cut.max) == F) {
+  if(is.null(cut.max) == FALSE) {
 
     dens.x.plot.i <- dens.x.i[dens.x.i < cut.max]
     dens.y.plot.i <- dens.y.i[dens.x.i < cut.max]
 
   }
+
+
+  if(cap.beans == TRUE) {
+
+    dens.x.plot.i <- dens.x.i[dens.x.i < max(dv.i) & dens.x.i > min(dv.i)]
+    dens.y.plot.i <- dens.y.i[dens.x.i < max(dv.i) & dens.x.i > min(dv.i)]
+
+  }
+
 }
 
 # BAR
@@ -1072,6 +1283,9 @@ points(x = rep(x.loc.i, length(dv.i)) + rnorm(length(dv.i), mean = 0, sd = jitte
 
 }
 
+if (quant.boxplot) {
+  quant <- c(0.25, 0.75)
+}
 # QUANTILES
 if (!is.null(quant)) {
 
@@ -1083,27 +1297,42 @@ if (!is.null(quant)) {
     quant.lwd <- c(rep(0.75, length(quant)))
   } else {quant.lwd <- rep(quant.lwd, length.out = length(quant))}
 
-  for (i in 1:length(quant)) {
-
-    # draw lines
-    segments(x.loc.i + (quant.length[i] - width.max), # left end
-             quantile(dv.i, probs = quant[i]),
-             x.loc.i - (quant.length[i] - width.max), # right end
-             quantile(dv.i, probs = quant[i]),
-             col =  colors.df$quant.col[bean.i],
-             lwd = quant.lwd[i],
-             lend = 3
-    )
+  if (quant.boxplot) {
+    for (i in 1:length(quant)) {
+      if (i == 1) {
+        segments(x.loc.i + (quant.length[i] - width.max),
+                 quantile(dv.i, probs = quant[i]) - (1.5*IQR(dv.i)),
+                 x.loc.i - (quant.length[i] - width.max),
+                 quantile(dv.i,probs = quant[i]) - (1.5*IQR(dv.i)),
+                 col = colors.df$quant.col[bean.i],
+                 lwd = quant.lwd[i], lend = 3)
+      } else {
+        segments(x.loc.i + (quant.length[i] - width.max),
+                 quantile(dv.i, probs = quant[i]) + (1.5*IQR(dv.i)),
+                 x.loc.i - (quant.length[i] - width.max),
+                 quantile(dv.i,probs = quant[i]) + (1.5*IQR(dv.i)),
+                 col = colors.df$quant.col[bean.i],
+                 lwd = quant.lwd[i], lend = 3)
+      }
+    }
+    segments(x.loc.i, quantile(dv.i, probs = min(quant)) - (1.5*IQR(dv.i)),
+             x.loc.i, quantile(dv.i, probs = max(quant)) + (1.5*IQR(dv.i)),
+             col = colors.df$quant.col[bean.i], lwd = quant.lwd[1],
+             lend = 3, lty = 1)
   }
-
-  # Vertical quant line
-
-  segments(x.loc.i,
-           quantile(dv.i, probs = min(quant)),
-           x.loc.i,
-           quantile(dv.i, probs = max(quant)), col = colors.df$quant.col[bean.i],
-           lwd = quant.lwd[1], lend = 3, lty = 1)
-
+  else {
+    for (i in 1:length(quant)) {
+      segments(x.loc.i + (quant.length[i] - width.max),
+               quantile(dv.i, probs = quant[i]),
+               x.loc.i - (quant.length[i] - width.max),
+               quantile(dv.i, probs = quant[i]), col = colors.df$quant.col[bean.i],
+               lwd = quant.lwd[i], lend = 3)
+    }
+    segments(x.loc.i, quantile(dv.i, probs = min(quant)),
+             x.loc.i, quantile(dv.i, probs = max(quant)),
+             col = colors.df$quant.col[bean.i], lwd = quant.lwd[1],
+             lend = 3, lty = 1)
+  }
 }
 
 # INFERENCE BAND
@@ -1122,11 +1351,33 @@ dens.inf.y <- dens.y.i[dens.x.i >= summary$inf.lb[bean.i] & dens.x.i <= summary$
 
 # Draw inf band
 
-if(inf.band == "wide") {
+if(inf.disp == "line") {
 
-rect(x.loc.i - width.max * .8,
+  segments(x.loc.i, summary$inf.lb[bean.i],
+           x.loc.i, summary$inf.ub[bean.i],
+           col = transparent(colors.df$inf.f.col[bean.i],
+                             trans.val = 1 - opac.df$inf.f.o[bean.i]),
+           lwd = inf.lwd[bean.i])
+
+  # Add whiskers
+
+  segments(rep(x.loc.i - width.max * .2, 2),
+           c(summary$inf.lb[bean.i], summary$inf.ub[bean.i]),
+           rep(x.loc.i + width.max * .2, 2),
+           c(summary$inf.lb[bean.i], summary$inf.ub[bean.i]),
+           col = transparent(colors.df$inf.f.col[bean.i],
+                             trans.val = 1 - opac.df$inf.f.o[bean.i]),
+           lwd = inf.lwd[bean.i])
+
+
+}
+
+
+if(inf.disp == "rect") {
+
+rect(x.loc.i - width.max * .7,
      summary$inf.lb[bean.i],
-     x.loc.i + width.max * .8,
+     x.loc.i + width.max * .7,
      summary$inf.ub[bean.i],
      col = transparent(colors.df$inf.f.col[bean.i],
                        trans.val = 1 - opac.df$inf.f.o[bean.i]),
@@ -1137,7 +1388,7 @@ rect(x.loc.i - width.max * .8,
 
 }
 
-if(inf.band == "tight") {
+if(inf.disp == "bean") {
 
   polygon(c(x.loc.i - dens.inf.y[1:(length(dens.inf.x))],
             x.loc.i + rev(dens.inf.y[1:(length(dens.inf.x))])),
@@ -1146,7 +1397,7 @@ if(inf.band == "tight") {
           col = transparent(colors.df$inf.f.col[bean.i],
                             trans.val = 1 - opac.df$inf.f.o[bean.i]),
           border = transparent(colors.df$inf.b.col[bean.i],
-                               trans.val = 1 - colors.df$inf.b.o[bean.i]),
+                               trans.val = 1 - opac.df$inf.b.o[bean.i]),
           lwd = bean.lwd[bean.i]
   )
 }
@@ -1159,7 +1410,7 @@ if(inf.band == "tight") {
 # AVERAGE LINE
 {
 
-  if(inf.band == "wide") {
+  if(inf.disp %in% c("line", "rect")) {
     segments(x0 = x.loc.i - width.max,
              y0 = summary$avg[bean.i],
              x1 = x.loc.i + width.max,
@@ -1171,7 +1422,7 @@ if(inf.band == "tight") {
     )
   }
 
-  if(inf.band == "tight") {
+  if(inf.disp == "bean") {
 
     fun.loc <- which(abs(dens.x.i - avg.line.fun(dv.i)) == min(abs(dens.x.i - avg.line.fun(dv.i))))
 
@@ -1185,6 +1436,7 @@ if(inf.band == "tight") {
              lend = 3
     )
   }
+
 
 }